requirements.txt
easygui<0.98.3
imageio<2.19.3
matplotlib<3.5.2
numpy<1.21.5
opencv_python<4.6.0.66
Pillow<9.3.0
streamlit<1.13.0
